Refactor right rail into DocumentationViewerControl using MVVM patterns
Extract documentation display functionality from MainWindow into a
reusable WPF UserControl following established MVVM patterns. The new
DocumentationViewerControl encapsulates all markdown rendering logic and
provides a clean interface through dependency properties and events.

Changes:
- Add DocumentationViewerControl with MVVM binding to MarkdownContent
- Add DocumentationViewerViewModel with self-contained documentation loading logic
- Remove 250+ lines of documentation code from MainWindow
- Replace inline markup with ItemsControl data binding
- Add DocumentId dependency property and DocumentChanged event
- Remove Quick Reference Links buttons
- Update all documentation loading calls to use new control

Benefits:
- Improved separation of concerns following existing UserControl patterns
- Self-contained documentation functionality
- Reusable component with clean interface
- Proper resource management with IDisposable
