+++ Controls/DocumentationViewerControl.xaml
@@ -1,0 +1,30 @@
+<UserControl x:Class="OneNoteAddinManager.Controls.DocumentationViewerControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="400">
+    <Grid>
+        <GroupBox Header="Context Help" Margin="5" Padding="10">
+            <Grid>
+                <!-- Documentation Content -->
+                <ScrollViewer 
+                              VerticalScrollBarVisibility="Auto" 
+                              HorizontalScrollBarVisibility="Disabled"
+                              Background="White"
+                              BorderBrush="LightGray"
+                              BorderThickness="1"
+                              Padding="10">
+                    <ItemsControl ItemsSource="{Binding MarkdownContent}">
+                        <ItemsControl.ItemsPanel>
+                            <ItemsPanelTemplate>
+                                <StackPanel/>
+                            </ItemsPanelTemplate>
+                        </ItemsControl.ItemsPanel>
+                    </ItemsControl>
+                </ScrollViewer>
+            </Grid>
+        </GroupBox>
+    </Grid>
+</UserControl>
