+++ MainWindow.xaml
@@ -333,52 +333,10 @@
                           Background="LightGray" 
                           ShowsPreview="True"/>
 
-            <!-- Right Rail - Context Help -->
-            <GroupBox Grid.Column="3" 
-                      Header="Context Help">
-                <Grid>
-                    <Grid.RowDefinitions>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="*"/>
-                    </Grid.RowDefinitions>
-                    
-                    <!-- Quick Reference Links -->
-                    <StackPanel Grid.Row="0" 
-                                Orientation="Horizontal" 
-                                Margin="0,0,0,10">
-                        <Button x:Name="OverviewLinkButton"
-                                Content="Overview"
-                                Click="OverviewLinkButton_Click"
-                                Padding="8,4"
-                                Margin="0,0,5,0"/>
-                        <Button x:Name="FieldsLinkButton"
-                                Content="Fields"
-                                Click="FieldsLinkButton_Click"
-                                Padding="8,4"
-                                Margin="0,0,5,0"/>
-                        <Button x:Name="TroubleshootingLinkButton"
-                                Content="Troubleshooting"
-                                Click="TroubleshootingLinkButton_Click"
-                                Padding="8,4"/>
-                    </StackPanel>
-                    
-                    <!-- Dynamic Markdown Viewer -->
-                    <ScrollViewer Grid.Row="1" 
-                                  VerticalScrollBarVisibility="Auto"
-                                  HorizontalScrollBarVisibility="Disabled"
-                                  Background="White"
-                                  BorderBrush="LightGray" 
-                                  BorderThickness="1">
-                        <StackPanel x:Name="MarkdownContent" 
-                                    Margin="15"
-                                    Background="White">
-                            <TextBlock Text="Loading context help..." 
-                                       FontFamily="Segoe UI" 
-                                       FontSize="12"/>
-                        </StackPanel>
-                    </ScrollViewer>
-                </Grid>
-            </GroupBox>
+            <!-- Right Rail - Documentation Viewer -->
+            <controls:DocumentationViewerControl x:Name="DocumentationViewer"
+                                                 Grid.Column="3"
+                                                 DocumentChanged="DocumentationViewer_DocumentChanged"/>
         </Grid>
 
         <!-- Status Bar -->
